Mode stuff to TRAMDataset. Fix some tests
