+++ deeptime/markov/msm/tram/__init__.py
@@ -1,2 +1,3 @@
 from ._tram import TRAM, unpack_input_tuple
 from ._tram_model import TRAMModel
+from ._tram_dataset import TRAMDataset
